Fixed Reverb
Thank you, Bomba, for encouraging me to look at the available pictures and schematics again to see if there was something I was missing. The reverb worked before, but it had to be driven hard for it to fill in with the wet-dry mix. After discovering how the real #002 reverb worked, the reverb fits in nicely and sits really well in the mix.